requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-google-genai
google-generativeai
python-multipart
python-dotenv
streamlit
requests
pandas
